# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 模型提供商管理服务 API 文档\n提供模型提供商、模型信息和参数规则的查询接口",
        "title": "Genkit AI Service API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/sessions": {
            "get": {
                "description": "获取用户的会话列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "获取会话列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否置顶",
                        "name": "isPinned",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否归档",
                        "name": "isArchived",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模型名称",
                        "name": "modelName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回会话列表",
                        "schema": {
                            "$ref": "#/definitions/SessionListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的聊天会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "创建新会话",
                "parameters": [
                    {
                        "description": "创建会话请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功创建会话",
                        "schema": {
                            "$ref": "#/definitions/SessionDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chat/sessions/search": {
            "get": {
                "description": "根据关键词搜索会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "搜索会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNo",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回搜索结果",
                        "schema": {
                            "$ref": "#/definitions/SessionListResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "参数验证失败",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/chat/sessions/{id}/archive": {
            "post": {
                "description": "归档或取消归档指定的会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "归档会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否归档",
                        "name": "archived",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功更新归档状态",
                        "schema": {
                            "$ref": "#/definitions/AnyDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权访问",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/chat/sessions/{id}/pin": {
            "post": {
                "description": "置顶或取消置顶指定的会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "置顶会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否置顶",
                        "name": "pinned",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功更新置顶状态",
                        "schema": {
                            "$ref": "#/definitions/AnyDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权访问",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AnyDataResponse": {
            "type": "object"
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "请求参数错误"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
        "SessionDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/SessionResponse"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "SessionListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "分页数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/PaginationData-array_model_SessionResponse"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 格式的 JWT 令牌进行身份认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}